services/dao-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
asyncpg==0.29.0
httpx==0.25.1
python-dotenv==1.0.0
nats-py==2.6.0

